refactor: renaming and moving some classes/packages
