+++ src/Tutor.Web/Startup.cs
@@ -10,8 +10,8 @@ using Microsoft.Net.Http.Headers;
 using System;
 using System.IO;
 using Tutor.Core.ContentModel;
+using Tutor.Core.DomainModel;
 using Tutor.Core.InstructorModel.Instructors;
-using Tutor.Core.KnowledgeComponentModel;
 using Tutor.Core.LearnerModel;
 using Tutor.Core.LearnerModel.Workspaces;
 using Tutor.Core.ProgressModel.Feedback;
